API/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.4.0
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.27
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0
tensorflow==2.15.0
tensorflow-hub==0.15.0
numpy==1.24.3
Pillow==10.2.0 